Improvements for the relative to absolute feature
`chdir()` does not work if multiple documents are opened from different
directories, because it sets the current directory for the whole process.
Pass the `base_path` as a parameter to the Lua script instead.

Also add an export option for this feature, and enable it for all formats
by default, which is especially useful for formats that contain internal
images.